EXPERIMENTAL: Remove seemingly unused code.
I couldn't find any tests or documentation that use this code branch, and the
purpose isn't clear to me (why would you pass in a field object as a value?).
It's a small performance gain to just remove it, but if anyone depends on it,
it could be added back.